operator: move all flags to a dedicated file
